+++ archive/grewtix/grewtix/asgi.py
@@ -4,7 +4,7 @@ ASGI config for grewtix project.
 It exposes the ASGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/3.2/howto/deployment/asgi/
+https://docs.djangoproject.com/en/3.0/howto/deployment/asgi/
 """
 
 import os
