+++ archive/grewtix/grewtix/urls.py
@@ -1,7 +1,7 @@
 """grewtix URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/3.2/topics/http/urls/
+    https://docs.djangoproject.com/en/3.0/topics/http/urls/
 Examples:
 Function views
     1. Add an import:  from my_app import views
@@ -17,6 +17,7 @@ from django.contrib import admin
 from django.urls import path, include
 
 urlpatterns = [
+    path('', include('tickets.urls')),
     path('admin/', admin.site.urls),
-    path('tickets/', include("tickets.urls")),
+    path('accounts/', include('django.contrib.auth.urls')),
 ]
