+++ archive/grewtix/grewtix/wsgi.py
@@ -4,7 +4,7 @@ WSGI config for grewtix project.
 It exposes the WSGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/3.2/howto/deployment/wsgi/
+https://docs.djangoproject.com/en/3.0/howto/deployment/wsgi/
 """
 
 import os
