+++ archive/grewtix/manage.py
@@ -5,7 +5,6 @@ import sys
 
 
 def main():
-    """Run administrative tasks."""
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'grewtix.settings')
     try:
         from django.core.management import execute_from_command_line
