+++ archive/grewtix/tickets/apps.py
@@ -2,5 +2,4 @@ from django.apps import AppConfig
 
 
 class TicketsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
     name = 'tickets'
