+++ grewtix/tickets/templates/ticket_index.html
@@ -1,0 +1,30 @@
+{% extends "header.html" %}
+{% load static %}
+{% block page_content %}
+<div class="container">
+    <br>
+    <div class="list-group">
+        <li class="list-group-item rectangle-link">
+            <a href="{% url 'tickets:unassigned_queue' %}"><button type="button" class="btn btn-primary">Unassigned
+                    Queue</button></a>
+        </li>
+        <li class="list-group-item rectangle-link">
+            <a href="{% url 'tickets:owner_queue' %}"><button type="button" class="btn btn-primary">My
+                    Queue</button></a>
+        </li>
+        <li class="list-group-item rectangle-link">
+            <a href="{% url 'tickets:created_queue' %}"><button type="button" class="btn btn-primary">Created By
+                    Me Queue</button></a>
+        </li>
+        <li class="list-group-item rectangle-link">
+            <a href="{% url 'tickets:recent_queue' %}"><button type="button" class="btn btn-primary">Recently
+                    Created Queue</button></a>
+        </li>
+        <li class="list-group-item rectangle-link">
+            <a href="{% url 'tickets:all_queue' %}"><button type="button" class="btn btn-primary">All Tickets
+                    Queue</button></a>
+        </li>
+    </div>
+    <br>
+</div>
+{% endblock %}
