+++ grewtix/tickets/templates/ticket_reports.html
@@ -1,0 +1,16 @@
+{% extends 'header.html' %}
+
+{% block page_content %}
+<div class="container">
+    This is the reports page..
+
+    <div class="row">
+        <div class="col-md-3">
+            <div class="card-body">
+                <h5 class="card-title">Total Tickets</h5>
+                <p class="card-text">This is where the # goes.</p>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
